+++ src/Commands/Commands.cpp
@@ -12,6 +12,7 @@
 #include "SaveVariablesToFile.h"
 #include "AutoLoad.h"
 #include "AutoGarrison.h"
+#include "AutoCapture.h"
 
 DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 {
@@ -23,6 +24,12 @@ DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 	MakeCommand<ToggleDesignatorRangeCommandClass>();
 	MakeCommand<AutoLoadCommandClass>();
 	MakeCommand<AutoGarrisonCommandClass>();
+	AutoCaptureCommandClass* autoCaptureT = GameCreate<AutoCaptureCommandClass>();
+	autoCaptureT->TechBuildingsOnly = true;
+	CommandClass::Array->AddItem(autoCaptureT);
+	AutoCaptureCommandClass* autoCaptureB = GameCreate<AutoCaptureCommandClass>();
+	autoCaptureB->TechBuildingsOnly = false;
+	CommandClass::Array->AddItem(autoCaptureB);
 
 	if (Phobos::Config::DevelopmentCommands)
 	{