+++ docs/Whats-New.md
@@ -496,6 +496,7 @@ New:
 - Custom exit cell for infantry factory (by Starkku)
 - Auto load hotkey command (by psi-cmd, Aephiex)
 - Auto garrison hotkey command (by Aephiex)
+- Auto capture hotkey command (by Aephiex)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)